add excludedFromCommands and move limited placing into misc
